requirements.txt
langgraph
langchain-community
langchain-google-genai
transformers
torch
torchvision
torchaudio
accelerate
opencv-python
scikit-learn
mediapipe
moviepy
ultralytics
concurrent.futures
multiprocessing
ray
asyncio
aiofiles
aiohttp
celery
redis
dask
distributed
gradio
streamlit
websockets
pandas
numpy
matplotlib
seaborn
plotly
ffmpeg-python
moviepy
google-generativeai
huggingface_hub
datasets
psutil
python-dotenv
pytube
